Bugs y otras mejoras.
- Solo hacer detach y remove si el entity manager contiene al objeto.
- Arreglado bug de que preguntas flashcards no puntuaban al
visualizarlas (darle al botón de continuar).
- Sólo mostramos enunciado y scrollbar en panel pregunta si la pregunta
tiene enunciado.
- Añadidos espacios entre palabra secreta de pregunta completar.

+++ java/Futbolistos/src/main/java/pds/futbolistos/bd/BaseDeDatos.java
@@ -103,7 +103,8 @@ public class BaseDeDatos {
 		iniciarTransaccion();
 		u.actualizarEstadisticas(sc);
 		u.removeSesion(sc);
-		em.remove(sc);
+		if (em.contains(sc))
+			em.remove(sc);
 		em.persist(u);
 		cerrarTransaccion();
 	}
